# docs/swagger.yaml
basePath: /api/v1
definitions:
  matter.CreateRequest:
    properties:
      content:
        type: string
      date:
        description: 设定的日期：2020-11-01
        type: string
      priority:
        description: 优先级，0/1/2/3 => 无优先级/低优先级/中优先级/高优先级
        type: integer
      state:
        description: 状态：0/1/2 => 未完成/已完成/已过期
        type: integer
      time:
        description: 设定的时间："12:02"
        type: string
      title:
        type: string
    type: object
  matter.FinishPayload:
    properties:
      state:
        description: 是否完成
        type: boolean
    type: object
  matter.GetResponse:
    properties:
      content:
        type: string
      date:
        description: 设定的日期：2020-11-01
        type: string
      priority:
        description: 优先级，0/1/2/3 => 无优先级/低优先级/中优先级/高优先级
        type: integer
      state:
        description: 状态：0/1/2 => 未完成/已完成/已过期
        type: integer
      time:
        description: 设定的时间："12:02"
        type: string
      title:
        type: string
    type: object
  matter.ListResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/matter.MatterInfo'
        type: array
    type: object
  matter.MatterInfo:
    properties:
      content:
        type: string
      date:
        description: 设定的日期：2020-11-01
        type: string
      priority:
        description: 优先级，0/1/2/3 => 无优先级/低优先级/中优先级/高优先级
        type: integer
      state:
        description: 状态：0/1/2 => 未完成/已完成/已过期
        type: integer
      time:
        description: 设定的时间："12:02"
        type: string
      title:
        type: string
    type: object
  matter.UpdateRequest:
    properties:
      content:
        type: string
      date:
        description: 设定的日期：2020-11-01
        type: string
      priority:
        description: 优先级，0/1/2/3 => 无优先级/低优先级/中优先级/高优先级
        type: integer
      state:
        description: 状态：0/1/2 => 未完成/已完成/已过期
        type: integer
      time:
        description: 设定的时间："12:02"
        type: string
      title:
        type: string
    type: object
  user.GetResponse:
    properties:
      nickname:
        type: string
      qq:
        type: string
      username:
        type: string
    type: object
  user.LoginRequest:
    properties:
      code:
        description: 小程序 auth code
        type: string
    type: object
  user.LoginResponse:
    properties:
      token:
        type: string
    type: object
  user.UpdateRequest:
    properties:
      nickname:
        type: string
    type: object
host: '...'
info:
  contact: {}
  license: {}
  title: todo-list
  version: "1.0"
paths:
  /matter:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/matter.ListResponse'
      summary: 获取所有事项
      tags:
      - matter
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/matter.CreateRequest'
          type: object
      responses:
        "200":
          description: OK
      summary: 创建事项
      tags:
      - matter
  /matter/{id}:
    delete:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: matter id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: 删除事项
      tags:
      - matter
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: matter id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/matter.GetResponse'
      summary: 查看事项详情
      tags:
      - matter
    put:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: matter id
        in: path
        name: id
        required: true
        type: string
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/matter.UpdateRequest'
          type: object
      responses:
        "200":
          description: OK
      summary: 修改事项
      tags:
      - matter
  /matter/{id}/finish:
    put:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: matter id
        in: path
        name: id
        required: true
        type: string
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/matter.FinishPayload'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/matter.FinishPayload'
      summary: 完成/取消完成事项
      tags:
      - matter
  /user:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.GetResponse'
      summary: 获取用户信息
      tags:
      - user
  /user/login:
    post:
      parameters:
      - description: body data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
          type: object
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.LoginResponse'
      summary: 登录
      tags:
      - user
swagger: "2.0"